+++ docs/CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove "type" field from Affect (OSIDB-2743)
 - Remove "type" field from Flaw (OSIDB-2735)
 - Remove "state" field from Flaw (OSIDB-2736)
+- Remove "resolution" field from Flaw (OSIDB-2737)
 
 ## [3.7.2] - 2024-05-17
 ### Fixed